add handler error all methods

+++ src/routes/deletePokemon.js
@@ -3,8 +3,12 @@ import { Pokemon } from "../db/sequelize.js";
 const deledePokemon = (app) => {
     app.delete('/api/pokemons/:id', (req, res) => {
         Pokemon.findByPk(req.params.id).then(pokemon => {
+            if(pokemon === null) {
+                const message = `Le pokemon demand' n'existe pas. Réessayez avec un autre identifiant.`
+                return res.status(404).json({message})
+            }
             const pokemonDeleted = pokemon;
-            Pokemon.destroy({
+            return Pokemon.destroy({
                 where: { id: pokemon.id }
             })
             .then(_ => {
